+++ Assets/Danmokou/Plugins/Danmokou/Danmaku/Expressions/TypedExpressionImpls.cs
@@ -171,7 +171,7 @@ public class TExGCX : TEx<GenCtx> {
         }
         if (t == tv2) return v2s.DictGetOrThrow(ExC(name), $"No v2 exists by name {name}.");
         if (t == tv3) return v3s.DictGetOrThrow(ExC(name), $"No v3 exists by name {name}.");
-        if (t == tvrv2) return rv2s.DictGetOrThrow(ExC(name), $"No V2RV2 exists by name {name}.");
+        if (t == tv2rv2) return rv2s.DictGetOrThrow(ExC(name), $"No V2RV2 exists by name {name}.");
         throw new Exception($"No handling in GenCtx for type {t.RName()}");
     }
 }